LayoutTests/media/track/track-load-error-readyState-expected.txt

Tests the error event on HTMLTrackElement and ERROR readyState on TextTrack.

EVENT(error)
EXPECTED (track.readyState == '3') OK
END OF TEST
